+++ packages/api-explorer/src/components/DocReferences/utils.tsx
@@ -25,56 +25,9 @@
  */
 import React from 'react'
 import type { ApiModel, IMethod, IType } from '@looker/sdk-codegen'
-import { Method, firstMethodRef } from '@looker/sdk-codegen'
 import { Link } from 'react-router-dom'
 import { RunItHeading } from '@looker/run-it'
-import { buildMethodPath, buildTypePath, highlightHTML } from '../../utils'
-
-/**
- * Returns the tag for a given method name
- * @param api Parsed api
- * @param methodName SDK method name
- * @returns Corresponding tag
- */
-const getMethodTag = (api: ApiModel, methodName: string) => {
-  // Find tag containing methodName
-  return Object.entries(api.tags)
-    .filter(([, methods]) => methodName in methods)
-    .map(([methodTag]) => methodTag)[0]
-}
-
-/**
- * Return the tag for a give type
- * @param api Parsed api
- * @param type to tag
- */
-const getTypeTag = (api: ApiModel, type: IType) => {
-  const method = firstMethodRef(api, type)
-  return getMethodTag(api, method.name)
-}
-
-/**
- * Builds a path matching MethodScene or TypeScene route
- * @param api parsed api
- * @param item A method or type item
- * @param specKey A string to identify the spec in the url
- * @returns a method or type path
- */
-export const buildPath = (
-  api: ApiModel,
-  item: IMethod | IType,
-  specKey: string
-) => {
-  let path
-  if (item instanceof Method) {
-    const tag = getMethodTag(api, item.name)
-    path = buildMethodPath(specKey, tag, item.name)
-  } else {
-    const tag = getTypeTag(api, item as IType)
-    path = buildTypePath(specKey, tag, item.name)
-  }
-  return path
-}
+import { buildPath, highlightHTML } from '../../utils'
 
 /**
  * Renders a heading and links to all item types
