+++ packages/api-explorer/src/utils/path.ts
@@ -24,6 +24,9 @@
 
  */
 
+import type { ApiModel, IMethod, IType } from '@looker/sdk-codegen'
+import { firstMethodRef, Method } from '@looker/sdk-codegen'
+
 /**
  * Builds a path matching the route used by MethodScene
  * @param specKey A string to identify the spec in the URL
@@ -49,3 +52,49 @@ export const buildTypePath = (specKey: string, tag: string, typeName: string) =>
 
 export const diffPath = 'diff'
 export const oAuthPath = 'oauth'
+
+/**
+ * Returns the tag for a given method name
+ * @param api Parsed api
+ * @param methodName SDK method name
+ * @returns Corresponding tag
+ */
+const getMethodTag = (api: ApiModel, methodName: string) => {
+  // Find tag containing methodName
+  return Object.entries(api.tags)
+    .filter(([, methods]) => methodName in methods)
+    .map(([methodTag]) => methodTag)[0]
+}
+
+/**
+ * Return the tag for a give type
+ * @param api Parsed api
+ * @param type to tag
+ */
+const getTypeTag = (api: ApiModel, type: IType) => {
+  const method = firstMethodRef(api, type)
+  return getMethodTag(api, method.name)
+}
+
+/**
+ * Builds a path matching MethodScene or TypeScene route
+ * @param api parsed api
+ * @param item A method or type item
+ * @param specKey A string to identify the spec in the url
+ * @returns a method or type path
+ */
+export const buildPath = (
+  api: ApiModel,
+  item: IMethod | IType,
+  specKey: string
+) => {
+  let path
+  if (item instanceof Method) {
+    const tag = getMethodTag(api, item.name)
+    path = buildMethodPath(specKey, tag, item.name)
+  } else {
+    const tag = getTypeTag(api, item as IType)
+    path = buildTypePath(specKey, tag, item.name)
+  }
+  return path
+}
